pyproject.toml
[build-system]
requires = ["pdm-backend>=2.0.0"]
build-backend = "pdm.backend"

[project]
name = "dao-review-system"
version = "0.1.0"
description = "DAO 任务复盘系统"
requires-python = ">=3.10"
dependencies = [
    "strawberry-graphql[fastapi]>=0.200.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "requests>=2.28.0",
    "schedule>=1.2.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]
lint = [
    "black>=24.0.0",
]
type = [
    "mypy>=1.8.0",
]

[project.scripts]
review-system = "src.review_system:main"

[tool.pdm.version]
source = "scm"
fallback-version = "0.0.0"

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
